Extract StatCard component
