+++ app/src/main/java/com/practica/proyectozoo/EspecieListActivity.kt
@@ -18,7 +18,6 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -158,34 +157,6 @@ fun EspecieListScreen(db: DatabaseHelper, modifier: Modifier = Modifier) {
     }
 }
 
-@Composable
-fun StatCard(
-    title: String,
-    value: String,
-    icon: ImageVector,
-    color: Color,
-    modifier: Modifier
-) {
-    Card(
-        modifier = modifier,
-        shape = RoundedCornerShape(12.dp),
-        colors = CardDefaults.cardColors(containerColor = color)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Column {
-                Text(title, fontSize = 14.sp, color = Color.DarkGray)
-                Text(value, fontSize = 22.sp, fontWeight = FontWeight.Bold)
-            }
-            Icon(icon, contentDescription = null, tint = Color.Black)
-        }
-    }
-}
 
 @Composable
 fun DropdownMenuEspecieItem(especie: Especie, db: DatabaseHelper, especies: MutableList<Especie>) {
